log-in and jwt

+++ StudySpace.Service/BusinessModel/RoomModel.cs
@@ -14,13 +14,13 @@ namespace StudySpace.Service.BusinessModel
 
         public int? Capacity { get; set; }
 
-        public decimal? PricePerHour { get; set; }
+        public double? PricePerHour { get; set; }
 
         public string Description { get; set; }
 
         public bool? Status { get; set; }
 
-        public decimal? Area { get; set; }
+        public double? Area { get; set; }
 
         public string Type { get; set; }
     }
